clean up plots, remove bug for spectral smoothing, edit readme, modify i
pre commit passed, except that isort wants an older python version?

Edited the scedc tutorial. we can keep it scedc specific because we don't yet have the option for the FDSN client.

I found a bug in the spectral normalization and added an if statement if we use it.

cleaned up a plotting module.

added text in readme, contributing, and modified citation.